Address PR 2695 feedback: Centralize prerelease policy and add boundary test
